Georgian language tests moved to separate file
